"10:10am. I got little sleep last night. It was one of those days where I lie in bed and instead of falling asleep start thinking even harder. But I did resolve some things. My regrets - though my rational mind says different, they all stem because I believe in outdated values of self improvement. To be more precise, though I figured out the self improvement loop, my rational mind is in conflict with what I wanted as a kid.
Eventually, I am going to overcome it and completely internalize this new way of thinking.

Rationally, there is no way a path to power would ever care about past failures or victories. It is all about the future. That I do care, is proof that I am half-normie and that my drive is half posturing. As a kid, I believed that a real man should always keep winning, but as I lived, I was constantly forced to make compromises in order to cultivate the actual path.

Me becoming a programmer was more of a desperation move than a premediated choice.

I do regret that being a 'real' man is not only emotional manipulation from the outside, but it is outright incompatible with a lot of requirements of the actual path. Trying to live up to wrong expectations just results in meaningless suffering.

I am good in this respect to some amount compared to other people, but I have weaknesses I have to resolve.

Ultimately, even I myself do not have a perfect image of how the post human should behave. I do not buy the rational ubermench image of characters like Fang Yuan completely. It is just too easy to have characters act a certain way and win in fiction.

10:25am. Let me chill a bit here and I will start.

Yesterday I tried out Rance Quest Magnum. At first I had issues with game screen being cut off near the bottom that I fixed by modifying the text scaling in the Windows settings, but after that I was surprised by how much fun I was having. Rance's chad antics put a constant grin on my face and made me laugh. Considering it is a plain jRPG at its core, I thought it would be a lot more boring than it actually is. It is not like I am unilaterally a Rance fan, I actually dropped Rance 4 due to it being boring. And Rance 2 wasn't that interesting even though I finished it. Kichikou and Sengoku are exceptions and they were both grand strategy RPGs.

10:45am. Maybe I did get some rest last night, even though throughout it felt I was awake the whole time.

I am not as tired as I expected. Let me start here.

10:50am. I've slacked for long enough. I'll finish today at 6pm so I can get some extra gaming time.

Until then let me just focus on the work in front of me.

10:55am. Let me get some thinking done. I need to decide what concrete steps I should take.

During the night, I've figured out how to parallelize the sample collection. Yesterday I said that CPSing is an absolute requirement, but I do not have enough. I am going to have to turn the game into an union type. Instead of having the game operate on the players, I will make it so that it returns the action nodes with a continuation. In essence I am going to compile the game to an union type.

This will make it easy to have an arbitrary number of instance run in parallel. Right now I can't do that. Even the CPS'd version that I have now does not return a continuation, instead for the human player it just calls the dispatcher.

I might have to redesign the whole thing to get this capability, but that is fine. If I did it like this, it would have the benefit of completely generalizing the other two ways.

11am. The reason why I'd want to do this is because GPUs suck for online learning. Running a single sample through the system would be 128 times slower than running 128 of them in parallel. And NN evaluation on the GPU is where the majority of the runtime overhead will be.

When I was running those PG in 2018 on a single sample, yes, I was doing something pretty stupid.

11:10am. Close down /g/ focus on the task at hand. It is time to get serious here. Though I feel really fatigued, I should give it my best shot for the day as usual. I can't give excuses and start slacking off because of things like my fatigue state.

...

Let me think. Yesterday I did the tabular CFR agent. Today I need to run it.

11:15am. Focus me. I already have the general outline.

What are the specific steps I have to take?

11:20am. The graph is troublesome as it requires a separate non-updating pass with just the average policy being used on both ends.

Nonetheless, let me go for it. Let me do a step of optimization, and a step of showing the value to me.

I should also try implementing linear CFR, since it has been mentioned to give two order of magnitude of speedup.

11:25am. Graph UI, graph UI... That is what I should focus on. Once I decided that I absolutely want to do something, things become easier.

Then instead of thinking what I want to do in addition to have to do things, I can just focus on the later. This gives me the ability to think more deeply about the subject. When one is one the fence, there is always the hesitation of expending too much mental energy on something that is fruitless.

I should do a number of training steps followed by a sampling step.

11:30am. Now I am thinking about GANs. Forget that, focus on the task at hand.

11:45am. Oh yeah, I was wrong about predictive coding being useless. I had an idea where those local targets could be used to make the replay buffer for each individual layer. That would also give me reward scaling invariance without any of the top level layer hacks.

But is more complicated and not as advantageous as my own scheme on the GPU.

12:25pm. I am thinking about that UI right now. I am narrowing down what my reqirements are. I am going to get rid of the kind of random replay buffer that I had in `ui_replay`.

Instead for trained tabular agents, I should just display the dictionary.

12:30pm. I am not going to bother with that step when it comes to flop poker.

https://gregoryszorc.com/blog/2021/04/06/surprisingly-slow/
https://www.thenewatlantis.com/publications/welcoming-our-new-robot-overlords

Let me stop here. The first article is interesting.

12:35pm. I want to think more. I am not eager to start just yet. After running the task through my mind enough times, I'll build up the motivation where I can't stop myself from starting.

One of the first things I will have to do is master some new widgets. For the next part, buttons and labels won't cut it. I am going to have to do some experimentation how to do tabs, and also those tree view like controls. I need to master showing and hiding widgets.

I've figured out how the recycle view works. Along with the charts, it is time to put those lessons to bear."