"10:15am. I am up. I was restful yesterday so I did not feel like going to bed early, but that ended up straining me. Right now I feel like I haven't gotten enough sleep.
10:20am. I keep iterating what I have to do in my head. Compared to yesterday, I've changed my mind a little.

I realized that the straightforward CFR is not the algorithm I want, instead I should first accumulate counterfactual rewards and then update the policy all at the end. I need to implement this algorithm and compare it to the original. Because this algorithm is closer in nature to the sampling one than the original.

10:30am. Huh, but if this is the right way of doing it, in the NN version, once the value nets absorb the values, how should the policy net samples be weighted?

11:20am. Ugh...I did not think about this problem at all.

It is easy enough in the tabular case to implement whatever algorithm I want. But in the NN version, I will be learning the values. After I learn them, how do I weight the policy net samples?

Suddenly Pure CFR seems a lot more attractive. Maybe I should have gone a step further with that old experiment, instead taken the max of all the values.

11:20am. If I could just learn the value nets and average them to get the policy that will make things a lot easier.

Let me take a look at that Phd I discovered by accident yesterday.

https://github.com/rggibson/open-pure-cfr
Regret Minimization In Games And The Development Of Champion Multiplayer Computer Poker-Playing Agents

Oh great. VS Code has the `Transform to Snake Case` command. Nice find!

11:30am. No, actually, I get how to get the policy values. Sure I might be learning the counterfactual values (divided by the player prob) but once I factor that out, they'll be regular values. Then I should scale the samples as usual.

1:05pm. I am brainstorming like crazy. Let me get breakfast.

2:20pm. Done with breakfast and chores. Let me finally read the thesis. My stress is sky high.

You know what, forget NNs. Forget them for a month or two.

I just realized - to get the baselines, I am not going to be able to train the NNs for the counterfactual values, but the real ones instead.

But just think what sort of problems this gives me.

To get the full equivalence, I'd need to train a NN for the average policy. Then a NN for the policy. Then a NN for the counterfactual values. Then a NN for the actual values.

Just thinking about it hurts me!

I mean, in the damn DREAM paper, they specifically don't train a NN for the avg policy, but average the past ones to lower the approximation error.

But now after this, I have 4 different NNs. Even if I ommit the average policy net, that is still 3.

3 nets who don't have any shared representation. Yes, I should share the representations, but how?

I hate NNs with a passion. I loathe them from the bottom of my heart!

The degree of my understanding is the degree of my connection with these algoriths. I made some progress with the buffer tricks, but that won't be enough.

Of course it is not enough.

Consider those predictive baselines. Since I have the full hidden state, I can just set them to the value directly. What is the analogue of that in the NN space.

In tabular regime, I can easily add and average values together. NNs require a whole compilation pass for the simplest of things. I just can't stand it. I am bashing my head against this wall, and am not getting much in return.

2:30pm. As far as algorithms go, CFR is wonderful. With some effort, I'll be able to make a breakthrough. But NNs...I should leave them out of my plans for now.

I need to focus on thoroughly mastering the tabular case. I'll find the inspiration that I need after that.

3pm. Today I am completely out of it. I am just wasting time.

80/159. This is so tough. Do I really need to be able to replicate these proofs to cause the Singularity? Just what do I have to do to understand this. It is like learning Chinese. Maybe I could do it if I put in the effort, but I know that I simply do not have the time or the energy to properly dedicate myself.

3:05pm. Forget it. Let me do some programming.

Yesterday it was light, but right now I am as stressed as I could be.

I am not in the mood for it at all. But let me just open the editor and copy the project. I'll do it all from scratch. Let me just measure the variance.

I made all sorts of plans, for example with a buffed up CFR algorithm, but now I am just digusted with that because I realized how much complexity this would add on the NN side.

My goals should be as follow:

Verify that the VR MCCFR from the paper works. My implementation has my own tricks, and because of that, clearly things are going wrong.

3:10pm. Based on my understanding, just the thing in the paper should work as well as regular CFR.

Maybe the way I am updating the average policy is wrong? This is one area where I am diverging from how it is done in the paper.

In the paper it is updated outright on line 3. That `(t-1) / t * s + 1 / t s'` is not a form I recognized before, but that thing with the ts is actually how a moving average is conputed inplace. During the night I remembered seeing it in the RL course by Litman and...that black dude.

3:20pm. I can't reason out anything. If I could do this in a braindead fashion like a code monkey, maybe I'd have been more successful.

3:30pm. Forget it. Let me take a nap. I am too angry, pissed and tired to do anything right now. My mindset is completely wrong."