"6pm. It is so painful. The pain of not understanding a piece of code is the biggest pain there is. Sigh. I know what I should be doing today, and yet I am overcome by my own limitations.
I want to reason as much as I can. I take pride in going more and more.

Logic is my greatest tool and revel in using it.

But I've reached my limit. I can't do it. I can't understand these ML algorithms by thinking about them. It hurts.

This is the same thing that killed me in high school.

6:05pm. To believe in logic is a curse.

As a kid I tried to reason out the universe and came to the conclussion that it would make more sense if nothing ever existed. Based on the whole that something cannot come out of nothing.

As an adult, in 2019 I made a rigorous study of formal proofs, and what did I discover? Logically, something cannot come out of nothing.

Whenever you have infinities, then the logic breaks down and anything goes.

Logical systems are big on termination checking.

So becoming better at reasoning, only served to deepen the chasm between myself and the universe itself.

6:10pm. I cannot come to an understanding between the view that universe cannot and yet does exist. I cannot mesh the two views. I cannot even imagine what sort of insight would be needed to cross this great chasm of understanding.

My demonic path tells me to assume that the logic is right.

I think this is where my life went wrong. This was my first real loss. When you internalize that you are a living inconsistency, a part of you starts to feel that it is alright to believe in anything.

I wish I lived in an universe whose existence made sense.

A part of me believes in solipsism, even though I don't know what it even is. Deep down I think that only God itself should be conscious being. A part of me believes that this would be justice.

But I have to live with the logical incoherency of me being conscious myself and having to deal with distressing reports of other people reporting being conscious as well.

6:15pm. A part of me due to solipsism as much as trend following believes that the Singularity should happen.

But it is April 2021, and only now am I starting work on sampling CFR. I can barely understand it. Do I really believe I can reach AGI with such a paltry level of understanding and intelligence? Yet, it must happen. Somehow. Just how?

6:20pm. Will years into the future the Singularity not have happened? Will I live out my whole life neither understanding myself nor the universe's existence?

Is the purpose of life the pursuit of happiness rather than power? Would it have been correct to live like an animal and trust my senses more?

6:20pm. For power, rob and cheat and steal...but can you actually steal undertanding the only thing of actual value? Can you cheat intelligence itself? Can you rob the skill of others?

Omnipotence is the greatest purpose, and yet the self improvement loop turns out to be iterative suicide. Necromutation is transcendence.

The demonic dao beckons.

Just how is my mind broken, that I do not understand all the most important things? That I have to live with these regrets? That I cannot summon inspiration as its master, yet live at its whims.

I can't imagine, even if I had a million times higher intelligence than now, how would I make progres on the universe existing problem.

6:30pm. This is my demonic path.

It makes me believe in logic, but confronts me with all the contradictions arising from that. It puts me in a universe that should not exist.

It makes me believe in the puruit of power, but then in a twist reveals that self improvement requires suicide to work. Just to get to it, I'd need to make a human level AI and make its power my own first. It does not tell me how.

It convinces me that both myself and others and animals and teaches me not to trust my senses, yet always in my moments of weakness like now when my reason reaches its end, I can only rely on that and be an animal.

6:35pm. Is this universe really not hell?

This pain I feel right now is surely the pain of not accepting my own limits. My intelligence is not enough, but my path does not allow me such compromises and excuses. My goal is too deeply ingrained to let something like this stop me.

I know deep down that I cannot have power without logic. Stories and superstition are not a substitute for code.

So what if in the future nothing comes to pass in the way that has been forecast. So what if I turn out to be some madman just pushing symbols around.

I cannot deal with the contradictions so the path itself should be responsible for them.

6:45pm. Using is fusing, I really should do some programming, but I am not in the mood right now. It is just too late to start anything. Tomorrow, I am going to forget what I've done so far and focus on getting those sample tests done. That will tell me whether VR MCCFR's variance reduction is working properly.

If not, that will be a huge gain as I will be able to make the agents work. If it does by some chance, things will be quite harder. I'll have to gradually move closer to what the paper says I should be doing."