"10:20am. I got up 10m ago. Let me chill for a while and then I will start. I feel like reading the PGM book instead of programming. But my determination to do both has gone up.
Compared to 2015 I really understand the book a lot more. Though it is just as droning as I remember it.

10:50am. Let me start for real. I want more knowledge. I want to do a complete tour of ML eventually.

11:05am. 90/1265. This is actually pretty boring. I am tempted to start skimming until I get to the algorithms.

...Yeah, let me do that. I'll never get anywhere with this book if I read it top to bottom. The crystalizations of knowledge are always algorithms. I should start with them and work my way from there.

https://www.youtube.com/watch?v=I7bgrZjoRhM
Nonparametric Bayesian Methods: Models, Algorithms, and Applications I

This is an interesting series. I see that Mike Jordan is a lecturer in part 2. Let me drop the book for now. It is too long and droning. These lectures seem more like what I want. Let me give the first a try.

11:35am. https://youtu.be/I7bgrZjoRhM?t=1173

I'll admit I did not know about how the Beta distribution worked before.

You know, an idea occurs to me. Now that those predictive coding papers showed me how to generate targets, doesn't that open an opportunity for analytic methods. I could use Bayesian methods instead of NNs...

12:35pm. I am getting some inspiration even though I won't quite be able to put it together until much later. It is not on using Bayesian methods. I am thinking about the difference between the prior and the sampling distribution.

Consider regularization, the way it is done is to put in some penalty term, but consider that in the light of my replay buffer ideas. Instead of changing the targets themselves, would it not make more sense for it to reweight the samples instead?

Yeah, that is it. I've always thought that there is something wrong for the weights to outright decay. Bayesian priors can be used as regularization terms to help shape the buffer.

https://www.youtube.com/watch?v=yfLoxwjCGNY
Nonparametric Bayesian Methods: Models, Algorithms, and Applications II

Mike said this would not be big on algorithms, but it is fine.

There is an itch I want to scratch. It feels like I am on the cusp of a qualitative breakthrough in understanding, but I need more knowledge and more math, and I have the motivation to learn it.

12:40pm. CFR is a great algorithm that seemingly comes out of the blue. Great algorithms always have a reason for their greatness. Given an algorithm I should be able to reason backwards to what the principles involved in its derivation are.

12:45pm. I want to know more of everything that works well, and bring the insights together.

Let me stop here so I can have breakfast."