"2:15pm. https://youtu.be/yfLoxwjCGNY?t=1319
He says that K means just flows out of this by setting the sigma to zero. I'd like to see how that could be done. He mentions it being done in his papers.

3:10pm. https://youtu.be/2H2n4iUYpZE
Nonparametric Bayesian Methods: Models, Algorithms, and Applications III

Let me take a break before I move on to this lecture.

https://github.com/AndrasKovacs/staged/blob/main/types2021/abstract.pdf

This got posted on the PL sub and I am skimming it.

3:20pm. I suppose this is in fact the right direction to develop staging in. Spiral is just a midpoint in the development of partial evaluation and having it top down would make things easier. But nonetheless, it would be too much work for me. I am not a PL researcher.

My attention is split between two fields, and my focus should have been on ML from the start.

...

Also, sigh. My mother got the latest results of analysis and it does not look good. It is some kind of cancer and she will have to go to chemotherapy. She seems fine now, but I can tell the reaper is getting closer. Will I have to experience her funeral a few years from now?

Things like this happen. Whether it is me or anybody else, the lack of control over what goes in the body, and the extreme complexity will kill us all. The results of nature's accumulation and development is what makes us, but it makes for a great hurdle to exceed.

If I had to face death myself, I'll do it, but I do not want it to be for such a stupid reason. Neither do I want this to happen to people I care about.

Nature is something to be exceeded as much as respect. You can respect it, but if it is going to kill you anyway, you might as well oppose it.

3:30pm. If I am being honest, causing the Singularity in the next 3 years, seems wildly unrealistic. At least for me. Maybe there is some genius in the world who could do it. I myself, the best I could do is get enough money to put her into cryosleep instead of a grave during this time.

Will I even be able to do this? It is not like I've been any successful in the past.

Maybe this experience will serve as a reminder not to slack off. I am going to finish these lectures today, and start work on the NN agent the first thing tomorrow. I can't let myself bear the endless wave of failure.

3:55pm. Done with chores. Let me resume.

Really, I haven't accomplished much in the last 6 years. Compared to the guy giving the talk, his insight is on an entirely different level than mine.

4pm. Focus me. Let me go for the second round.

4:20pm. https://youtu.be/2H2n4iUYpZE?t=1295

He mentions Arron Cinlar's book 'Probability and Stochastics' here. He mentions it is focused on Poisson distributions rather than Guassians.

https://www.springer.com/gp/book/9780387878584

Ah it is Ehran Cinlar. Ok, I'll get the book from Libgen. I am in such a mood.

4:25pm. The math he is showing there, yes, it is beyond me, but it does not seem as far as it would be usually. Usually, I'd be completely flummoxed at what the point is, but I can sort of appreciate it now.

Maybe all my efforts did have some point to them.

4:30pm. Forget the book for now, I'll leave it in the folder. Let me get back to the lecture.

5:05pm. https://youtu.be/2H2n4iUYpZE?t=3095

I am not sure I am paying attention to this anymore, but thankfully the video is almost over. I'll take a look at the next one as well. I also want to watch the belief propagation one. This is what Youtube search got me when I put in Bayesian message passing. This is one the algorithms I implemented in the PGM course, but never quite internalized much like the rest of them. I want to revisit it. I know that PGM book has algorithms in it. I might try skimming it tomorrow.

Damn, are these lectures droning. Some integral here and there as he talks about some abstract thing.

My programming ability while strong in human relative terms is just so insignificant in the grand scheme of things. Without the self improvement loop, I'll attain and hold on to nothing.

I am relying that the community at large comes up with some better algorithms than those that exist now.

https://youtu.be/EUUyQbtUXR0
Nonparametric Bayesian Methods: Models, Algorithms, and Applications IV

I really hope this has some actual algorithmics. So far the title of the series really does not deserve the algorithm part. It barely had anything on them.

Let me continue. I really am sleep right now.

https://www.youtube.com/watch?v=Xs2QIUAzB4I
Where The Really Hard Problems Really Are?

The title of this caught my attention. I'll take a look at it later.

5:40pm. https://news.ycombinator.com/item?id=26907176

I am reading the Cerberas thread from yesterday instead of watching the video. Let me go for the video for 5 more minutes. At this point, the lectures have exceeded my boredom threeshold and I am running on fumes.

No forget that video. This is how my math excursion usually go. Instead of math lectures, what I really want is algorithms I can think about and obsess over. Let me take a look at the belief propagation video.

https://www.youtube.com/watch?v=meBWAboEWQk
Belief Propagation

6:05pm. https://youtu.be/meBWAboEWQk?t=893

This kind of hierarchy is much more biologically plausible than what backprop does. I do want to study this for a while.

I do want a tour of the algorithms once again. I implemented this during the PGM course, but could not get much out of it due to insufficient understanding at the time. Right now I want to at least get the basics stored in the back of my mind.

https://www.youtube.com/watch?v=VE2ITg_hGnI
On Gradient-Based Optimization: Accelerated, Distributed, Asynchronous and Stochastic

This is a talk by Mike Jordan. I wonder what he has to say on the subject. Not expected him to talk about this.

https://www.youtube.com/watch?v=mJGwqhbwqjA
Elchanan Mossel (MIT): Simplicity and Complexity of Belief Propagation

I'll also leave this talk for tomorrow.

Right now my fatigue is over the limit. I do not have the will for yet another lecture.

https://www.youtube.com/watch?v=Xs2QIUAzB4I
Where The Really Hard Problems Really Are?

Let me also this again so I can take a look at it.

6:20pm.

///

>Time for an AI winter
with AI chips burning 15KW? No chance for the winter in sight. Some chances for AI hell though.

///

Oh lol. From that Cerberas HN thread.

Let me stop here. Maybe I'll watch a bit on the Belief Prop video on the sly if I feel like it. But probably not. Rest is important.

Right now, there is only a week left in a month. I'll spend a few days studying Bayesian algorithms, just enough to satisfy me and then I am going to seriously get on work on finishing that poker player and getting on from that stage of my life. Once the computer is doing most of the work instead of me, I know I will have made it."