"12:30pm. Let me go to bed here. I am not ready to start after all. But I'll leave out fiction and frivolous pursuits in return. Until I am ready, I'll stay away from the screen.
7:10pm. Up to a bit over 6pm, I spent all the time since the last entry in bed. The fear is giving way to great sadness which is has abated by half, but I am still not ready to start. I think that one more day of this should be enough to give me the kick that I need.

I did spend some time thinking about programming as well.

And it is harsh. It is really difficult to adapt enumerative methods to the GPU. It might be possible in theory, but the reasoning gets hard really quickly. I do not want to go down this path. Instead, the outcome sampling methods are the ones that are easily tractable in parallel. I can thing of a few different variants. I am not sure which I will end up with in the end, but I'll start off with the straightforward one, with no TD learning or cutoffs.

It is a lot easier to start off with a fixed number of particles and filter them as things go along.

With enumerative methods, I have no idea how I would deal with chance nodes.

I really did have the intention of doing enumerative CFR at least for Leduc even one the GPU, but I do not want to bother with it.

It is true that the first time I implemented sampling CFR I had difficulty getting it to work, but that was because of all the mistakes that I won't repeat. It really isn't that much more difficult than the vanilla.

7:30pm. I have the methods that I need, but if only I can get my mood to improve I should be able to start. The shock treatment away from the screen should be the most effective. Eventually I'll get tired of moping around.

To start things off, I am first going to implement sampling CFR's parallel version in the tabular regime. After that I'll focus on dealing with NNs.

7:35pm. Sigh, let me get back to RI. Once I catch up, I will finally be free.

7:40pm. I'll go to bed early today as well, so let me chill a bit here."