"9:30am. I go up 10m ago and am chilling. Let me do that for a bit longer and then I will start.
9:50am. Let me start.

10am.

```
inl policy player_state game_state pid actions =
    (log_prob_from_sample (1 / f64 (length actions)), sampling.sample actions), fun r => r
```

How do I make a batch version of the uniform player? Also I just now realized that pids can be different between samples here!

That means, I can't use anything but self play in parallel training.

`player_state game_state pid actions`, these 4 args should be in an array. And the output should be an array of `(log_prob_from_sample (1 / f64 (length actions)), sampling.sample actions), fun r => r`. No way around this.

It is a pity I can only do self play training here, but this will have the effect of making things even simpler for me.

...Though there is an option to partition the data by the ids. That would allow me to use two separate players.

Let me take a short break here.

10:25am. Let me resume.

10:30am. I am thinking of how I will deal with `r2`'s in the neural net. Instead of indexing into it, I will just give it as it is.

For both the policy net, I think I'll just make the pid a feature. Otherwise in HU NL Holdem, I won't be able to deal with cutoffs properly.

10:35am. I said it was to contribute to my own side, but my internal story is changing again.

I am simplifying my image. In truth programming does not matter for its own sake.

I am imaging picking a real world goal, and imagining how having an AI agent guide me would help.

Imagine I was playing chess against Magnus Carlsen for a large sum of money. If it was just me, I'd be quite nervous, but if I had AlphaZero giving me tips on the side, I'd be a lot more confident.

The same would go with poker.

Rather than automating everything off the bat, I'll train the agent, and play it for a while myself with him as the advisor. The same would go if I were cultivating the arts. I should not rush through this.

Right now it is May 2021, and I am nothing more than a beginner. I bet plenty of people could have picked my own path in 2015, but not have gotten caught up in doing random courses and then a ML library and a programming langauge. Had they done that, they'd be in a far stronger position than I am now.

Spiral will pay for itself long term, but for now forget those accomplishments.

I am like a millionaire who has lost it all. Forget the past luxuries. It is time to build a new empire.

10:50pm. Programming is the kind of activity where having high pride gives great benefits. But cultivating agents might be better served with humility.

Pride...I am such a natural channeler of it. I started off good at programming, and then I never compromised even once in the quest for omnipotence despite all the hardship.

I never actually saw a point to humility. Some people think it is a virtue, but I discarded that as a cultural thing with no basis in reality.

Still, I can see clearly that pride is the wrong thing to channel in the face of ML's winds of uncertainty.

The demands on programming brilliance are not high. I do not think that even excellence in mathematics is the determining factor here, since we are all stealing insights from everybody else.

Rather than demonically trying to squeeze out the insight from the aether, the challenge here is to give yourself to the machine.

I want to be great. I want to be king. But the game here is to be a beggar and a philantropist. For the sake of your own side that is.

11:10am. I need to be more flexible. Rigid trees cannot resist the wind.

I am thinking...

Obviously, the uniform player is not the one that will give me trouble.

But I am thinking about the linear player. I haven't expected the pid thing to be an issue. I've thought about the serialization in the outputs.

But should I partition the game by pids? Could I do it with just a single player?

11:30am. No. Things are a lot more complicated than I originally envisioned.

What I envisioned was a set number of threads that terminate at different times, but that I am thinking about it more in depth I see that the situation is not only that, but that pids could potentially be different.

Suppose I used a single player for both the players and put in the pid as a feature - that might give me trouble when start using RNNs. I do not want to commit myself to feedforward nets completely just yet.

11:35am. Hmmmm, no if I did restrict myself to a single player, that would make things simpler.

11:50am. This is so complicated. Every little architecture change that I can think of, from going to feedforeward to RNNS to transformers would require a change in the training regime.

11:55am. This is rough on me. This is the same thing as me not being able to do training for tabular players.

Well, for tabular players, it is not like it specifically can't be done.

12pm. Input filtering is rough.

It would be more efficient to add some TD learning to the mix.

12:20pm. Agh, let me take a break here. It is time for that anyway.

The problem isn't really implementing the specific thing I want here, right now.

The problem is that for every different learning algorithm and NN family, I would need to implement a different training regime.

Just like with going from single trace to parallel right now, things feel off for me. I know how to do it, but I am having trouble finding abstractions opportunities. The patterns are too complex for my functional magic to work. It seems I'd be better off to do an unique implementation for everything.

12:25pm. I have only 30 chapters left and then I will finally be free from Reverend Insanity. I thought it had 2400 something, but I am only 30 chapters away from finishing it.

https://www.youtube.com/watch?v=CY_LEa9xQtg
Risto Miikkulainen: Neuroevolution and Evolutionary Computation | Lex Fridman Podcast

I want to watch this interview since it is on evo computation.

...

I think it is fine if I spend my time thinking about the programming problem in front of me before committing.

If I have to implement a different training function for every architecture and learning algo, so be it. But I should not jump into it before I feel that this is the case.

There is a lot of work here, but there isn't a huge unbounded amount of it like when I worked on an ML library and the Spiral language. All this will go quickly once I catch my bearings."